Emin branch

+++ src/main/java/org/example/domain/dao/FlightDao.java
@@ -4,6 +4,6 @@ import org.example.domain.entity.FlightEntity;
 
 
 public abstract class FlightDao
-        implements Dao<FlightEntity,Integer> {
+        implements Dao<FlightEntity, Long> {
 
 }
